Update pom.xml, CustomConditionTest.java, and 2 more files...

+++ src/test/java/org/brit/FileDownLoadTest.java
@@ -1,8 +1,11 @@
 package org.brit;
 
+import com.codeborne.selenide.Condition;
 import com.codeborne.selenide.Configuration;
 import com.codeborne.selenide.FileDownloadMode;
+import com.codeborne.selenide.WebDriverRunner;
 import org.apache.commons.io.FileUtils;
+import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
 
 import java.io.File;
@@ -16,14 +19,25 @@ import static org.openqa.selenium.By.linkText;
  * @author sbrit
  */
 public class FileDownLoadTest {
-    @Test
-    public void testTestWithProxy() throws IOException {
+    @BeforeClass
+    public void beforeFileDownLoadTestClass() throws IOException {
         Configuration.downloadsFolder = "downloads";
         Configuration.proxyEnabled = true;
         Configuration.fileDownload = FileDownloadMode.PROXY;
+        open("http://the-internet.herokuapp.com/upload");
+        $("#file-upload").uploadFromClasspath("fileToDownload.txt");
+        $("#file-submit").click();
+        $("#uploaded-files").shouldBe(Condition.visible, Condition.exactText("fileToDownload.txt"));
+        if (new File(Configuration.downloadsFolder).exists()){
+            FileUtils.cleanDirectory(new File(Configuration.downloadsFolder));
+        }
+    }
+
+
+    @Test
+    public void testTestWithProxy() throws IOException {
         open("https://codepen.io/gladchinda/full/GemGNG");
         switchTo().frame("result");
-        FileUtils.cleanDirectory(new File(Configuration.downloadsFolder));
         $(linkText("Export Photos")).click();
         sleep(1500);
         List<String> strings = FileUtils.readLines(new File(Configuration.downloadsFolder + "/photos.csv"), "UTF-8");
@@ -32,9 +46,7 @@ public class FileDownLoadTest {
 
     @Test
     public void testTestWithoutProxy() throws IOException {
-        Configuration.downloadsFolder = "downloads";
         open("http://the-internet.herokuapp.com/download");
-        FileUtils.cleanDirectory(new File(Configuration.downloadsFolder));
         File file = $(linkText("UploadNotepad.txt")).download();
         List<String> strings = FileUtils.readLines(file, "UTF-8");
         strings.forEach(System.out::println);
@@ -42,11 +54,7 @@ public class FileDownLoadTest {
 
     @Test
     public void testTestWithProxy2() throws IOException {
-        Configuration.downloadsFolder = "downloads";
-        Configuration.proxyEnabled = true;
-        Configuration.fileDownload = FileDownloadMode.PROXY;
         open("http://www.jtricks.com/bits/content_disposition.html");
-        FileUtils.cleanDirectory(new File(Configuration.downloadsFolder));
         File file = $(linkText("Text file with Content-Type of text/plain.")).download();
         List<String> strings = FileUtils.readLines(file, "UTF-8");
         strings.forEach(System.out::println);
